Changed short to int in the execution loop

+++ src/nazario/emulator/util/FlagFunction.java
@@ -5,5 +5,5 @@ import nazario.emulator.Registrars;
 
 @FunctionalInterface
 public interface FlagFunction {
-    void apply(short programCounter, Memory memory, Registrars registrars, int size, int cycles);
+    void apply(int programCounter, Memory memory, Registrars registrars, int size, int cycles);
 }
